+++ cypress/support/component.tsx
@@ -1,6 +1,6 @@
 import { mount } from "cypress/react18";
-import AppContextProvider from "@/Components/molecules/providers/AppContextProvider";
 import "cypress-real-events";
+import AppContextProvider from "@/app";
 
 Cypress.Commands.add("mount", (component, options) => {
     return mount(<AppContextProvider>{component}</AppContextProvider>, options);
